+++ RELEASE_NOTES.md
@@ -8,5 +8,6 @@ This is the first release of Pixel Fortress ! ✨
 - Add CI to build artifacts and publish them on GitHub
 - Combat units are exploring the map to uncover fog of war until they find an enemy
 - Add a minimap displaying the world
+- Add an option to display health bar on damaged units
 
 Thank you for playing!